[ABI] Don't emit overriding associated type declarations into witness tables.
When an associated type defined in a given protocol has the same name as
an associated type in an inherited protocol, it is always equivalent to
the "overridden" associated type. Don't emit a new entry in the
witness table for the overriding associated type.

Saves ~68k of binary size in the standard library.