Removed some TF attr modifiers (e.g. $dtype) (#19009)
* Removed all TF attr modifiers (e.g. $dtype) except for $in and $tensor.

Also extended attr sanity check code in TFDeabstraction::formGraphOp() to make
sure for a $tensor attr, its # scalar values match the # of scalars specified in
the subsequent shape attr. e.g. We have 2 scalar elements below.

```swift
  let shape : TensorShape = [1, 2]
  _ = #tfop("Const", dtype: Float.self, value$tensor: [17.0 as Float, 18.0], shape: shape)
```

* Refined error msg.

* Added back the $shape attr modifier based on feedback from @lattner -- when we
see an int array, we need that modifier for compiler to infer it's a shape attr.

* Fixed unit tests by removing $dtype.