Sema: Remove special handling of single 'Any' parameter
This was an artifact of some other Swift 3 compatibility behavior that has
since been removed.

There's a weird behavior change here. We would previously solve the
constraint if the LHS was a type variable and RHS is Any, which is
unsound, but removing this code changes something else in the solver
which causes the test for SR-2505 in test/Constraints/closures.swift
to become ambiguous.

I hacked around this by adding a double penalty for converting function
types to Any.