Sema: Check for type variable first in matchExistentialTypes()
Even if the RHS is 'Any', we don't want to solve the constraint if
'LHS' is a type variable, because the type variable might be bound
to a noescape function type later.

Noticed by inspection and I don't have a test case; it may well
be NFC because of how CSBindings works.