Fixed some memory allocation issues in const eval (#19040)
* Before this PR, ConstExprEvaluator has its own allocator. When
`constantEvaluator` is destroyed in `TFDeabstractionPass::run()`, the allocated
memory is returned, and that causes some symbolic values to be overwritten in
subsequent compiler passes due to new allocation, causing TFPartition pass to
read invalid symbolic values (as tfop attrs).

One example is this test case:
```swift
public func test() {
  let _: ResourceHandle = #tfop("AnonymousIterator",
                                output_types: [Float.self],
                                output_shapes: [nil as TensorShape?,
                                                TensorShape([1])])
}
```

Right after const eval in deabstraction, the 2 shapes in `output_shapes` have
symbolic values:

```
  enum: case none
  enum: case some(Wrapped), payload:   agg: 1 elt:     array<Int32>: 1 elt:       agg: 1 elt:         inst:   %16 = integer_literal $Builtin.Int32, 1
```

But in TFPartition, those values have been stomped on, causing
`hostFn.print(*outs)` in `TFFunctionPartition::markFunction()` to fail.

The fix is to use the allocator associated with the ASTContext of the SIL
module, whose lifetime covers both the deabstractio and partition passes.

With this fix, we can continue to use RK_Inst-typed symbolic values, even when
the underlying `inst` is dead-code-removed. In the above example, the second
shape element (of type Optional<TensorShape>) contains a const int backed by
inst `%16 = integer_literal $Builtin.Int32, 1`, but that inst is removed (by
performance pipeline) by the time TFPartition gets run. The inst is however
still accessible in TFPartition, since it's bump-pointer allocated, and the
allocator lifetime as described above covers the TFPartition pass.

* Brought back the fix for using RK_Inst typed const values.