Remove SourceManager::getBufferIdentifierForLoc
And tidy up doc comments for other methods that do or don't respect
'#sourceLocation'.

'getBufferIdentifierForLoc' is no longer useful: it doesn't
consistently return either the name of an actual buffer or the name of
a file suitable for diagnostics. As seen in the previous commit, all
remaining uses of it were wrong anyway. Remove it.