[Sema][Diag] Provide diagnostics for invalid 'for?' use
Amend the 'must be unwrapped' error when a optional sequence is used in a for-in loop.
The additional and preferred fixit becomes 'use for?'. This is a problem since we
don't have a way to propagate the stmt to CSDiagnostics.cpp, where this diagnostic is emitted.
But we don't really need to: in the case of for-in loops, this error can be diagnosed outside
of the constraint system. Currently, it is diagnosed in 'performStmtDiagnostics' in
MiscDiagnostics.cpp. However, I had to ultimalety #include ConstraintSystem.h to be able to call
the fixit methods and avoid duplicating logic.