Demangling: use `LLVM_BUILTIN_UNREACHABLE` over `llvm_unreachable`
Avoid a dependency on LLVMSupport at runtime through the `llvm_unreachable`.
This would pull in `llvm_unreachable_internal` in debug builds, which requires a
runtime dependency on LLVMSupport which increases the size of the binary
considerably.  Although we miss the error messages in debug builds, we would
normally lose the error message on a release build so this doesn't make it any
worse.