SIL: Don't implode parameters into a tuple in TypeConverter::getLoweredFormalTypes()
getLoweredFormalTypes() behaved differently for top-level
functions and methods:

- For functions it would bridge each parameter type and return a new
  function type with the bridged parameters.

- For methods, it would build a new function type whose argument list
  was always a tuple of the form ((Args...), Self), with all formal
  parameters of the method imploded into a single tuple argument.

This is a simple change in isolation, but AbstractionPatterns for
Objective-C methods and C functions imported as methods depended on
the old representation.

This patch reworks AbstractionPattern:getFunctionParamType() to do
the right thing for these AbstractionPattern kinds.

Note that I haven't touched the older getFunctionInputType() which
still does the old thing, but it's going away.