SESE Loops: transform loops to have single exit block in non-cloning case.  (#19385)
Algorithm Outline
(Note that exiting blocks are inside the loop and exit blocks are outside the loop. Therefore, an exit edge is from an exiting block to an exit block.)

This PR takes care of transforming the loop such that the common post dominator of all exit blocks is the new exit block for the loop. This is one of the several transformations needed to eliminate undef from SESE loops. Consider the following snippet:

while(...) {
   if (...) {
     ... 
     break;
   }
}
Recall that the blocks within if do not belong to the while loop in the SIL IR. The transformation implemented in this PR has the effect of moving the blocks back into the loop.

Step 1. Transform Loop. (See ensureSingleExitBlock)

Let nearestCommonPD be the nearest common post-dominator block of all the exit blocks.
Move all the blocks reachable from exiting blocks to the nearestCommonPD inside the loop. (It is not always possible to move all reachable blocks without cloning some of the them. See notes below.)
Let exitBlocks be the set of exit blocks for the loop after the above transformation.

Step 2. Connect exiting blocks to a new latch and unify loop arguments. (See patchEdges) (Only changes related to exit arguments are shown here.)

Let exitArgs be the union of the arguments of all exitBlocks.
for each exitArg in exitArgs, add an appropriate argument to the new header block. Let p0, p1, ..pn be the new arguments corresponding to exit arguments a0, a1, a2, ..., an.
For a exit edge br exit_i(x, y) -> exit_i(a2, a3), change the source to br new_latch(p0, p1, x, y, p4, ..., pn)

Notes. If a block is not dominated by the header of the loop, we cannot move it into the loop in Step 1. For such cases, we will need to clone the block before moving it into the loop. I will send out a separate PR for that purpose.