build: use the target variables for the ICU check
The ICU library is used for the target library, not the host tools.  Use the
appropriate variable `SWIFT_PRIMARY_VARIANT_SDK` and
`SWIFT_PRIMARY_VARIANT_ARCH` to select the right CMake variables for the ICU
definitions.  There is no need to use the `_default` suffixed versions as the
defaults have already been used to set the value above.