build: redcue the computation of variant suffixes and directories
We currently locally compute the variant suffix and directories all across the
build system.  The computation sometimes is incorrect (like when building
non-MachO targets).  Reduce the number of places that we compute this by using a
helper function through out.  The remaining places that use
`SWIFT_SDK_<SDK>_LIB_SUBDIR` do so as an alias for the sdk variant lowercased
rather than as a means of getting the subdir.  Once those are reduced, the
location where the subdir variable is used should more or less be localised to
the utilities and will permit us to remove the variable entirely which will make
it easier to follow where the files are destined to go.