[sil] Introduce FullApplySiteKind and ApplySiteKind to allow for exhaustive switching over FullApplySites and ApplySiteKind.
Currently there is a bug in the closure specializer that was caused by
BeginApply not being handled correctly. Rather than just fixing that and leaving
the badness, I am instead in this commit introducing enums for apply sites so we
can avoid this problem in the future by using exhaustive switches to guide
developers adding new types of apply sites in the future.

rdar://44612356